removed the implementation in the api module
