+++ cmd/metrics-scraper/app/db/database.go
@@ -1,22 +1,22 @@
 package db
 
-import (
-	"database/sql"
-	"log"
-)
+// import (
+// 	"database/sql"
+// 	"log"
+// )
 
-var (
-	db *sql.DB
-)
+// var (
+// 	db *sql.DB
+// )
 
-func init() {
-	// Create database connection
-	var err error
-	db, err = sql.Open("sqlite", "app_sync.db")
-	if err != nil {
-		log.Fatalf("Error opening app_sync database: %v", err)
-	}
-}
+// func init() {
+// 	// Create database connection
+// 	var err error
+// 	db, err = sql.Open("sqlite", "app_sync.db")
+// 	if err != nil {
+// 		log.Fatalf("Error opening app_sync database: %v", err)
+// 	}
+// }
 
 // func GetDB() *sql.DB {
 // 	return db
